data: Move mime-type from shared-mime-info to here
So that the thumbnailer is still functional, but without adding
a necessary stat and magic check for each and every PNG file around.

This is particularly a problem for the icon cache, as each PNG file
in a theme would end up being checked for which one of the variants
it was.